refactor(app/pages): 🚚 update imports of the sections with the new paths
